feat: redesign details page
