+++ assets/scss/variables.scss
@@ -1,3 +1,5 @@
+$font-size-bigtitle: 32px;
+$font-size-title: 24px;
 $font-size-subtitle: 20px;
 $font-size-large: 16px;
 $font-size-normal: 14px;
@@ -6,9 +8,9 @@ $font-size-smaller: 8px;
 $font-weight-semibold: 600;
 $font-weight-black: 900;
 
-$color-primary: #053CC7;
+$color-primary: #008CFF;
 $color-primary-60: #F6F6F6;
-$color-secondary: #5E85E6;
+$color-secondary: #82C4FA;
 $color-font-100: #121212;
 $color-font-60: #A4A4A4;
 $color-font-10: #FFFFFF;
