+++ components/auctions/details/Description.vue
@@ -1,19 +1,19 @@
 <template>
-  <v-col cols="12" class="mt-4 mb-5">
-    <h3 class="font-weight-bold text-primary text-subtitle-1">Descrição</h3>
-    <p class="text-justify text-body-2 opacity-50">
+  <v-col cols="12" class="mt-2 mb-5">
+    <div class="font-weight-semibold text-primary">Descrição</div>
+    <div class="text-justify text-font-60 mt-2 description">
       {{ showMore ? description : truncatedDescription }}
       <span
         v-if="description.length > maxChars"
         variant
         small
         color="primary"
-        class="open-text"
+        class="cursor-pointer text-secondary text-decoration-underline"
         @click="toggleShowMore"
       >
         {{ showMore ? "Ver menos" : "Ver mais" }}
       </span>
-    </p>
+    </div>
   </v-col>
 </template>
 <script setup lang="ts">
@@ -34,10 +34,9 @@ const toggleShowMore = () => {
   showMore.value = !showMore.value;
 };
 </script>
+
 <style scoped>
-.open-text {
-  cursor: pointer;
-  color: #1976d2;
-  text-decoration: underline;
+.description {
+  line-height: 20px !important;
 }
 </style>
