--- components/layouts/default/MainCard.vue
@@ -1,5 +1,0 @@
-<template>
-  <v-card class="px-8 pt-2 rounded-lg h-100" elevation="4">
-    <slot />
-  </v-card>
-</template>
