+++ layouts/default.vue
@@ -4,7 +4,7 @@
     <LayoutsDefaultAppBar />
     <v-main>
       <v-container
-        class="d-flex justify-center align-center px-8 pa-0 bg-transparent"
+        class="d-flex justify-center align-center px-8 pa-0 bg-primary-60"
         fluid
       >
         <slot />
@@ -15,21 +15,8 @@
 
 <style scoped>
 :deep(.v-application__wrap) {
-  min-height: 0;
-}
-
-:deep(.v-application) {
-  background-color: transparent !important;
-}
-
-:deep(.v-container) {
-  width: 100%;
-  height: 100%;
-  margin-left: 0;
-  margin-right: 0;
-  background-image: url("@/assets/images/background.svg");
-  background-size: cover;
-  background-position: center;
-  background-repeat: no-repeat;
+  min-height: calc(100vh - 64px);
+  padding-bottom: 12px;
+  background-color: rgba(var(--v-theme-primary-60)) !important;
 }
 </style>
